Refactor SIMD module: Modularize ops, add arch-specific implementations
