+++ CHECKLIST_CURRENT.md
@@ -1,0 +1,55 @@
+# CFD Suite - Current Status Checklist
+
+## Build Status ✅
+- [x] Core library builds without errors
+- [x] HDF5 properly optional
+- [x] SIMD module restructured (<500 lines)
+- [x] GPU infrastructure initialized
+
+## Critical Issues ❌
+- [ ] Field2D missing `set` method
+- [ ] Examples don't compile (API mismatch)
+- [ ] GPU dispatch not connected
+- [ ] Integration tests failing
+
+## Architecture Improvements ✅
+- [x] Modular structure enforced
+- [x] No redundant implementations
+- [x] Proper feature gating
+- [x] Clean naming (no adjectives)
+
+## SIMD/GPU Status
+### SIMD ✅
+- [x] AVX2/SSE4.2 (x86_64)
+- [x] NEON (AArch64)
+- [x] SWAR fallback
+- [ ] Missing: sum_f32, max_f32, add_u32
+
+### GPU ⚠️
+- [x] WGPU context creation
+- [x] 4 compute shaders (WGSL)
+- [ ] Dispatch integration
+- [ ] Performance validation
+
+## Solver Status
+### Working ✅
+- [x] SIMPLE/PISO
+- [x] CG/BiCGSTAB
+- [x] Discretization schemes
+- [x] Time integration
+
+### Incomplete ⚠️
+- [ ] LBM (no streaming)
+- [ ] Turbulence validation
+- [ ] Multiphase reconstruction
+- [ ] Unstructured mesh
+
+## Next Priority Actions
+1. Fix Field2D API
+2. Complete GPU dispatch
+3. Fix example compilation
+4. Validate algorithms
+5. Complete documentation
+
+## Honest Assessment
+**Alpha Quality**: Solid architecture, working SIMD, incomplete features. Requires ~40% more work for production readiness.
